requirements.txt
# Core dependencies with explicit versions
pydantic==2.5.2
httpx==0.24.1
python-dotenv==1.0.0
pydantic-ai==0.0.19

# YNAB API integration
requests==2.31.0

# Utility libraries
python-dateutil==2.8.2

# CLI interface
click==8.1.7

# Streamlit interface
streamlit==1.29.0

# Additional dependencies from original requirements
google-generativeai==0.8.4
PyYAML==6.0.1
structlog
pandas
numpy
pytest
mypy
black
isort
scikit-learn
google-api-python-client==2.123.0
google-auth==2.28.2
typing==3.7.4.3
typing-extensions==4.10.0 